refactor: Move networking functions to gallia.net
