+++ tests/pytest/test_net_if.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from gallia.utils import net_if_addrs
+from gallia.net import net_if_addrs
 
 
 def test_net_if_addrs() -> None:
